pyproject.toml
[tool.poetry]
name = "cookies-test"
version = "0.1.0"
description = ""
authors = ["iizhukov <iizhukov@icloud.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.112.0"
uvicorn = "^0.30.6"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
server="main:main"
